Rename to lower name for ESLint Unicorn.
