+++ vscode/src/status-bar.ts
